+++ .github/release-notes.md
@@ -1,3 +1,4 @@
 * Show self-inflicted deaths in killfeed
 * Fix error when steam path is incorrect
 * Store past player avatars
+* Show note contents in tooltip